automation-system/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pyyaml==6.0.1
aiohttp==3.9.1
asyncio-mqtt==0.16.1

# Database & Storage
supabase==2.3.0
psycopg2-binary==2.9.9

# Notion Integration
notion-client==2.2.1

# Task Scheduling
apscheduler==3.10.4
celery==5.3.4

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
